+++ src/InlinePanel.php
@@ -2,10 +2,9 @@
 
 namespace KirschbaumDevelopment\NovaChartjs;
 
-use App\Nova\Resource;
 use Laravel\Nova\Panel;
 use Illuminate\Http\Request;
-use Illuminate\Support\Facades\App;
+use App\Nova\Resource as NovaResource;
 use KirschbaumDevelopment\NovaChartjs\Contracts\Chartable;
 
 class InlinePanel extends Panel
@@ -13,17 +12,18 @@ class InlinePanel extends Panel
     /**
      * Create a new panel instance containing NovaChartjsChart.
      *
-     * @param Resource $resource
+     * @param NovaResource $resource
      * @param Request $request
      * @param string $panelTitle
      * @param bool $showLabel
      * @param bool $notEditable
+     * @param bool $hideFromIndex
      */
-    public function __construct(Resource $resource, Request $request, $panelTitle = 'Chart Metric Values', $showLabel = false, $notEditable = false)
+    public function __construct(NovaResource $resource, Request $request, $panelTitle = 'Chart Metric Values', $showLabel = false, $notEditable = false, $hideFromIndex = false)
     {
         parent::__construct(
             $panelTitle,
-            $this->prepareFields($this->fields($resource->resource, $request, $panelTitle, $showLabel, $notEditable))
+            $this->prepareFields($this->fields($resource->resource, $request, $panelTitle, $showLabel, $notEditable, $hideFromIndex))
         );
     }
 
@@ -35,14 +35,15 @@ class InlinePanel extends Panel
      * @param mixed $panelTitle
      * @param bool $showLabel
      * @param bool $notEditable
+     * @param bool $hideFromIndex
      *
      * @return array
      */
-    protected function fields(Chartable $chartable, Request $request, $panelTitle = 'Chart Metric Values', $showLabel = false, $notEditable = false): array
+    protected function fields(Chartable $chartable, Request $request, $panelTitle = 'Chart Metric Values', $showLabel = false, $notEditable = false, $hideFromIndex = false): array
     {
         $field = NovaChartjs::make($panelTitle, 'novaChartjsMetricValue', function () use ($chartable) {
             return $chartable->novaChartjsMetricValue->metric_values ?? [];
-        })->chartable($chartable ?? App::make($request->viaResource()::$model));
+        });
 
         if ($showLabel) {
             $field->showLabel();
@@ -50,6 +51,10 @@ class InlinePanel extends Panel
 
         if ($notEditable) {
             $field->notEditable();
+        }
+
+        if ($hideFromIndex) {
+            $field->hideFromIndex();
         }
 
         return [