+++ src/Traits/HasChart.php
@@ -7,8 +7,11 @@ use KirschbaumDevelopment\NovaChartjs\Models\NovaChartjsMetricValue;
 
 trait HasChart
 {
+    /** @var array */
+    protected $unsavedMetricValues = [];
+
     /**
-     * Get the Chartable Model's metric values
+     * Get the Chartable Model's metric values.
      *
      * @return MorphOne
      */
@@ -18,27 +21,36 @@ trait HasChart
     }
 
     /**
-     * Delete a models chart data before model is deleted
+     * Delete a models chart data before model is deleted.
      */
-    public static function bootHasNovaChartjsChart()
+    public static function bootHasChart()
     {
         static::deleting(function ($model) {
             if ($model->novaChartjsMetricValue) {
                 $model->novaChartjsMetricValue->delete();
             }
         });
+
+        static::created(function ($model) {
+            if (! empty($model->unsavedMetricValues)) {
+                $model->novaChartjsMetricValue()->create(['metric_values' => $model->unsavedMetricValues]);
+            }
+        });
     }
 
     /**
-     * Mutator to set Metric Values from Chartable model
+     * Mutator to set Metric Values from Chartable model.
      *
      * @param $value
      */
     public function setNovaChartjsMetricValueAttribute($value): void
     {
         if (! $this->novaChartjsMetricValue) {
-            $novaChartjsMetricValue = new NovaChartjsMetricValue(['metric_values' => $value]);
-            $this->novaChartjsMetricValue()->save($novaChartjsMetricValue);
+            if ($this->getKey()) {
+                $this->novaChartjsMetricValue()->create(['metric_values' => $value]);
+            } else {
+                $this->unsavedMetricValues = $value;
+            }
         } else {
             $this->novaChartjsMetricValue->metric_values = $value;
             $this->novaChartjsMetricValue->save();
@@ -46,9 +58,9 @@ trait HasChart
     }
 
     /**
-     * Return a list of all models available for comparison to root model
+     * Return a list of all models available for comparison to root model.
      *
-     * @return \Illuminate\Database\Eloquent\Collection
+     * @return array
      */
     public static function getNovaChartjsComparisonData(): array
     {
@@ -59,7 +71,7 @@ trait HasChart
     }
 
     /**
-     * Return a list of additional datasets added to chart
+     * Return a list of additional datasets added to chart.
      *
      * @return array
      */