+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # ChangeLog
 This document is meant for tracking notable changes to `nova-chartjs`
 
-## ChangeList
-Coming Soon
+## 0.1.0 
+1. Chart metric values can now directly be added while creating new Chartable model. 
+2. Deprecated `RedirectsOnCreate` Trait
+
+# 0.2.0
+1. Deprecated `chartable` method on `NovaChartjs`
+2. Added `hideFromIndex` option to `InlinePanel`